+++ app/controllers/shoppe/countries_controller.rb
@@ -1,6 +1,5 @@
 module Shoppe
   class CountriesController < Shoppe::ApplicationController
-
     before_filter { @active_nav = :countries }
     before_filter { params[:id] && @country = Shoppe::Country.find(params[:id]) }
 
@@ -15,9 +14,9 @@ module Shoppe
     def create
       @country = Shoppe::Country.new(safe_params)
       if @country.save
-        redirect_to :countries, :flash => {:notice => t('shoppe.countries.create_notice')}
+        redirect_to :countries, flash: { notice: t('shoppe.countries.create_notice') }
       else
-        render :action => "new"
+        render action: 'new'
       end
     end
 
@@ -26,15 +25,15 @@ module Shoppe
 
     def update
       if @country.update(safe_params)
-        redirect_to [:edit, @country], :flash => {:notice => t('shoppe.countries.update_notice') }
+        redirect_to [:edit, @country], flash: { notice: t('shoppe.countries.update_notice') }
       else
-        render :action => "edit"
+        render action: 'edit'
       end
     end
 
     def destroy
       @country.destroy
-      redirect_to :countries, :flash => {:notice => t('shoppe.countries.destroy_notice')}
+      redirect_to :countries, flash: { notice: t('shoppe.countries.destroy_notice') }
     end
 
     private
@@ -42,6 +41,5 @@ module Shoppe
     def safe_params
       params[:country].permit(:name, :code2, :code3, :continent, :tld, :currency, :eu_member)
     end
-
   end
 end