+++ app/controllers/shoppe/delivery_services_controller.rb
@@ -1,6 +1,5 @@
 module Shoppe
   class DeliveryServicesController < Shoppe::ApplicationController
-
     before_filter { @active_nav = :delivery_services }
     before_filter { params[:id] && @delivery_service = Shoppe::DeliveryService.find(params[:id]) }
 
@@ -15,9 +14,9 @@ module Shoppe
     def create
       @delivery_service = Shoppe::DeliveryService.new(safe_params)
       if @delivery_service.save
-        redirect_to :delivery_services, :flash => {:notice => t('shoppe.delivery_services.create_notice')}
+        redirect_to :delivery_services, flash: { notice: t('shoppe.delivery_services.create_notice') }
       else
-        render :action => "new"
+        render action: 'new'
       end
     end
 
@@ -26,15 +25,15 @@ module Shoppe
 
     def update
       if @delivery_service.update(safe_params)
-        redirect_to [:edit, @delivery_service], :flash => {:notice => t('shoppe.delivery_services.update_notice')}
+        redirect_to [:edit, @delivery_service], flash: { notice: t('shoppe.delivery_services.update_notice') }
       else
-        render :action => "edit"
+        render action: 'edit'
       end
     end
 
     def destroy
       @delivery_service.destroy
-      redirect_to :delivery_services, :flash => {:notice =>t('shoppe.delivery_services.destroy_notice')}
+      redirect_to :delivery_services, flash: { notice: t('shoppe.delivery_services.destroy_notice') }
     end
 
     private
@@ -42,6 +41,5 @@ module Shoppe
     def safe_params
       params[:delivery_service].permit(:name, :code, :active, :default, :courier, :tracking_url)
     end
-
   end
 end