+++ app/controllers/shoppe/product_categories_controller.rb
@@ -1,6 +1,5 @@
 module Shoppe
   class ProductCategoriesController < Shoppe::ApplicationController
-
     before_filter { @active_nav = :product_categories }
     before_filter { params[:id] && @product_category = Shoppe::ProductCategory.find(params[:id]) }
 
@@ -15,9 +14,9 @@ module Shoppe
     def create
       @product_category = Shoppe::ProductCategory.new(safe_params)
       if @product_category.save
-        redirect_to :product_categories, :flash => { :notice => t('shoppe.product_category.create_notice') }
+        redirect_to :product_categories, flash: { notice: t('shoppe.product_category.create_notice') }
       else
-        render :action => "new"
+        render action: 'new'
       end
     end
 
@@ -26,22 +25,22 @@ module Shoppe
 
     def update
       if @product_category.update(safe_params)
-        redirect_to [:edit, @product_category], :flash => { :notice => t('shoppe.product_category.update_notice') }
+        redirect_to [:edit, @product_category], flash: { notice: t('shoppe.product_category.update_notice') }
       else
-        render :action => "edit"
+        render action: 'edit'
       end
     end
 
     def destroy
       @product_category.destroy
-      redirect_to :product_categories, :flash => { :notice => t('shoppe.product_category.destroy_notice') }
+      redirect_to :product_categories, flash: { notice: t('shoppe.product_category.destroy_notice') }
     end
 
     private
 
     def safe_params
-      params[:product_category].permit(:name, :permalink, :description, :image_file, :parent_id, :permalink_includes_ancestors)
+      file_params = [:file, :parent_id, :role, :parent_type, file: []]
+      params[:product_category].permit(:name, :permalink, :description, :parent_id, :permalink_includes_ancestors, attachments: [image: file_params])
     end
-
   end
 end