+++ app/models/shoppe/order.rb
@@ -1,5 +1,7 @@
 module Shoppe
   class Order < ActiveRecord::Base
+    EMAIL_REGEX = /\A\b[A-Z0-9\.\_\%\-\+]+@(?:[A-Z0-9\-]+\.)+[A-Z]{2,6}\b\z/i
+    PHONE_REGEX = /\A[+?\d\ \-x\(\)]{7,}\z/
 
     self.table_name = 'shoppe_orders'
 
@@ -13,25 +15,25 @@ module Shoppe
     require_dependency 'shoppe/order/delivery'
 
     # All items which make up this order
-    has_many :order_items, :dependent => :destroy, :class_name => 'Shoppe::OrderItem', :inverse_of => :order
-    accepts_nested_attributes_for :order_items, :allow_destroy => true, :reject_if => Proc.new { |a| a['ordered_item_id'].blank? }
+    has_many :order_items, dependent: :destroy, class_name: 'Shoppe::OrderItem', inverse_of: :order
+    accepts_nested_attributes_for :order_items, allow_destroy: true, reject_if: proc { |a| a['ordered_item_id'].blank? }
 
     # All products which are part of this order (accessed through the items)
-    has_many :products, :through => :order_items, :class_name => 'Shoppe::Product', :source => :ordered_item, :source_type => 'Shoppe::Product'
+    has_many :products, through: :order_items, class_name: 'Shoppe::Product', source: :ordered_item, source_type: 'Shoppe::Product'
 
     # The order can belong to a customer
-    belongs_to :customer, :class_name => 'Shoppe::Customer'
-    has_many :addresses, :through => :customers, :class_name => 'Shoppe::Address'
+    belongs_to :customer, class_name: 'Shoppe::Customer'
+    has_many :addresses, through: :customers, class_name: 'Shoppe::Address'
 
     # Validations
-    validates :token, :presence => true
-    with_options :if => Proc.new { |o| !o.building? } do |order|
-      order.validates :email_address, :format => {:with => /\A\b[A-Z0-9\.\_\%\-\+]+@(?:[A-Z0-9\-]+\.)+[A-Z]{2,6}\b\z/i}
-      order.validates :phone_number, :format => {:with => /\A[\d\ \-x\(\)]{7,}\z/}
+    validates :token, presence: true
+    with_options if: proc { |o| !o.building? } do |order|
+      order.validates :email_address, format: { with: EMAIL_REGEX }
+      order.validates :phone_number, format: { with: PHONE_REGEX }
     end
 
     # Set some defaults
-    before_validation { self.token = SecureRandom.uuid  if self.token.blank? }
+    before_validation { self.token = SecureRandom.uuid  if token.blank? }
 
     # Some methods for setting the billing & delivery addresses
     attr_accessor :save_addresses, :billing_address_id, :delivery_address_id
@@ -48,8 +50,8 @@ module Shoppe
     #
     # @return [Float] - the length of time
     def build_time
-      return nil if self.received_at.blank?
-      self.created_at - self.received_at
+      return nil if received_at.blank?
+      created_at - received_at
     end
 
     # The name of the customer in the format of "Company (First Last)" or if they don't have
@@ -85,16 +87,15 @@ module Shoppe
     #
     # @return [Integer]
     def total_items
-      order_items.inject(0) { |t,i| t + i.quantity }
+      order_items.inject(0) { |t, i| t + i.quantity }
     end
 
-    def self.ransackable_attributes(auth_object = nil)
-      ["id", "billing_postcode", "billing_address1", "billing_address2", "billing_address3", "billing_address4", "first_name", "last_name", "company", "email_address", "phone_number", "consignment_number", "status", "received_at"] + _ransackers.keys
+    def self.ransackable_attributes(_auth_object = nil)
+      %w(id billing_postcode billing_address1 billing_address2 billing_address3 billing_address4 first_name last_name company email_address phone_number consignment_number status received_at) + _ransackers.keys
     end
 
-    def self.ransackable_associations(auth_object = nil)
+    def self.ransackable_associations(_auth_object = nil)
       []
     end
-
   end
 end