+++ app/models/shoppe/setting.rb
@@ -2,17 +2,16 @@ require 'ostruct'
 
 module Shoppe
   class Setting < ActiveRecord::Base
-    
     # Validations
-    validates :key, :presence => true, :uniqueness => true
-    validates :value, :presence => true
-    validates :value_type, :presence => true
-    
+    validates :key, presence: true, uniqueness: true
+    validates :value, presence: true
+    validates :value_type, presence: true
+
     before_validation do
-      self.value_type = I18n.t("shoppe.settings.types")[self.key.to_sym].try(:capitalize) || self.value.class.to_s
+      self.value_type = I18n.t('shoppe.settings.types')[key.to_sym].try(:capitalize) || value.class.to_s
       self.value      = encoded_value
     end
-    
+
     # The encoded value for saving in the backend (as a string)
     #
     # @return [String]
@@ -23,7 +22,7 @@ module Shoppe
       else                        value.to_s
       end
     end
-    
+
     # The decoded value for the setting attribute (in it's native type)
     #
     # @return [Object]
@@ -36,7 +35,7 @@ module Shoppe
       else                        value.to_s
       end
     end
-    
+
     # A full hash of all settings available in the current scope
     #
     # @return [Hash]
@@ -46,23 +45,22 @@ module Shoppe
         h
       end
     end
-    
-    # Update settings from a given hash and persist them. Accepts a 
+
+    # Update settings from a given hash and persist them. Accepts a
     # hash of keys (which should be strings).
     #
     # @return [Hash]
     def self.update_from_hash(hash)
-      existing_settings = self.all.to_a
+      existing_settings = all.to_a
       hash.each do |key, value|
-        existing = existing_settings.select { |s| s.key.to_s == key.to_s }.first
+        existing = existing_settings.find { |s| s.key.to_s == key.to_s }
         if existing
-          value.blank? ? existing.destroy! : existing.update!(:value => value)
+          value.blank? ? existing.destroy! : existing.update!(value: value)
         else
-          value.blank? ? nil : self.create!(:key => key, :value => value)
+          value.blank? ? nil : create!(key: key, value: value)
         end
       end
       hash
     end
-    
   end
 end