+++ db/migrate/20141026181717_allow_multiple_shoppe_products_per_shoppe_product_category.rb
@@ -11,17 +11,13 @@ class AllowMultipleShoppeProductsPerShoppeProductCategory < ActiveRecord::Migrat
     # define the old belongs_to association (as it's no longer on the model)
     Shoppe::Product.class_eval do
       belongs_to :old_category,
-                 :class_name => "Shoppe::ProductCategory",
-                 :foreign_key => "product_category_id"
+                 class_name: 'Shoppe::ProductCategory',
+                 foreign_key: 'product_category_id'
     end
-    # migrate over to our new join table. With a rather hachy rescue if there is something wrong with the products category to start with
+    # migrate over to our new join table
     Shoppe::Product.all.each do |product|
-      begin
-        product.old_category ? product.product_categories << product.old_category : "Something wrong with #{product.inspect}".red
-        product.save
-      rescue Exception => e
-        puts "#{product.inspect}: #{e.message}".red
-      end
+      product.product_categories << product.old_category
+      product.save
     end
     # lastly, remove the old product_category_id and associated index
     remove_index :shoppe_products, :product_category_id if index_exists?(:shoppe_products, :product_category_id)
@@ -35,8 +31,8 @@ class AllowMultipleShoppeProductsPerShoppeProductCategory < ActiveRecord::Migrat
     # define the old belongs_to association once again as we're going to re-add our goodies
     Shoppe::Product.class_eval do
       belongs_to :new_category,
-                 :class_name => "Shoppe::ProductCategory",
-                 :foreign_key => "product_category_id"
+                 class_name: 'Shoppe::ProductCategory',
+                 foreign_key: 'product_category_id'
     end
     # migrate over from the new table to the old association
     Shoppe::Product.all.each do |product|